Add Custom Dashboard for Super admin, store admin and store assistant, Restricted some routes based on user role, Update custom menu items
